Finished first revision of signaling.
E.g.

try'
    signal: "stuff"
Catch' |aCtx aSignal|
    (resume: aCtx With: "Resumed with some value")

yields

"Resumed with some value"

since tryCatch is an expression

The use og signal: "stuff" forwards control to the handler in catch, with then resumes (sends back the control) with a supplied value to the try body, which then in return yields the supplied value ("Resumed with some value").